move middlewares from main to other files
